Refactored functionality from Sphere into Primitive.
Moved material accesses into Primitive.  Also refactored get_normal to
correctly calculate normal for Sphere and modified get_color_contribution
to use get_normal instead of calculating directly.  This prepares us to
move get_color_contribution to Primitive as this logic is common to all
renderable objects.

A suggested refactoring was to keep Primitive as a pure interface and 
to put the common code into a subclass of Primitive.  This option is not
exercised on this commit but remains a future option.



git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@14 a272ff57-b438-0410-9c92-d377374fc898