This commit refactors all floats into doubles.  It also moves get_color_contribution to Primitive.
As a consequence of moving get_color_contribution to Primitive, get_center was also moved to Primitive.
This was done to eliminate the need for Primitive to have knowledge of it's subclass Sphere, if that
would even be possible.  The consequence of this is that all primitives now supported the concept of
having a center.  This may prove necessary later anyway (without a center, where does the object
get rendered at?). 




git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@22 a272ff57-b438-0410-9c92-d377374fc898