Checkboard material added.
Since .so loading isn't up yet, I've directly added plugins/checkerboard.o to 
the link line, so that will need to be replaced.

Also, there is some issues with the checkermaterial itself.  Something
wrong in the calcuation that is keeping it from properly checkering.

Finally, I kicked the subpixels down to squareof=2 while developing for
test_reflection.xml.  I'd leave it up at 5 but this Dell Celeron is leaving
scorch marks on the coffee table everytime I render :-]




git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@40 a272ff57-b438-0410-9c92-d377374fc898