Noise works.  It doesn't feel quite right; i may need to tweak it a bit. Also added material attribute parsing, which meant one more Material contructor. Changed makefile to use 'xv' instead of 'cygstart'.  This should make it more platform independent.
NOTE: There is a nasty bug around reflections with the noise material.  I haven't isolated it yet;
it doesn't appear to be in noisematerial.cpp but there may be some assumptions that I made that I 
need to validate.  Essentially, if the reflection and reflectivity are not equal in the two materials,
you get a wierd band of white followed by blown out colors facing the light source. If they are
equal, at any value, it works correctly, although I haven't noticed much difference in reflection
unless they are zero'ed out.



git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@43 a272ff57-b438-0410-9c92-d377374fc898